try catch was added on api call, context was modified
