+++ src/components/AppProvider/AppProvider.tsx
@@ -1,30 +1,32 @@
 import React, { useState } from 'react';
+import { getMovie } from '../../api';
 import { Movie } from '../../types/Movie';
+import { MovieData } from '../../types/MovieData';
 
 type Context = {
   movies: Movie[],
-  setMovies: (newMovies: Movie[]) => void,
   query: string,
-  setQuery: (newQuery: string) => void,
   movie: Movie | null,
-  setMovie: (newMovie: Movie | null) => void,
   isMovieExist: boolean,
-  setIsMovieExist: (newCheck: boolean) => void,
   isSearch: boolean,
-  setIsSearch: (newStatus: boolean) => void,
+
+  handleInput: (event: React.ChangeEvent<HTMLInputElement>) => void,
+  getMovieFromServer: (appliedQuery: string) => void,
+  handleSearch: (event: React.FormEvent<HTMLFormElement>) => void,
+  addingMovie: () => void,
 };
 
 export const AppContext = React.createContext<Context>({
   movies: [],
-  setMovies: () => {},
   query: '',
-  setQuery: () => {},
   movie: null,
-  setMovie: () => {},
   isMovieExist: false,
-  setIsMovieExist: () => {},
   isSearch: false,
-  setIsSearch: () => {},
+
+  handleInput: () => {},
+  getMovieFromServer: () => {},
+  handleSearch: () => {},
+  addingMovie: () => {},
 });
 
 export const AppProvider: React.FC = ({ children }) => {
@@ -34,17 +36,85 @@ export const AppProvider: React.FC = ({ children }) => {
   const [isMovieExist, setIsMovieExist] = useState(false);
   const [isSearch, setIsSearch] = useState(false);
 
+  const handleInput = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const newValue = event.target.value;
+
+    setQuery(newValue);
+
+    if (newValue === '' || newValue !== query) {
+      setIsMovieExist(false);
+    }
+  };
+
+  const getMovieFromServer = async (appliedQuery: string) => {
+    try {
+      const newFilm = await getMovie(appliedQuery);
+
+      setIsSearch(false);
+
+      if ('Poster' in newFilm) {
+        const {
+          Poster,
+          Title,
+          Plot,
+          imdbID,
+        } = newFilm as MovieData;
+
+        const movieImg = Poster === 'N/A'
+          ? 'https://via.placeholder.com/360x270.png?text=no%20preview'
+          : Poster;
+
+        const newMovieInList = {
+          title: Title,
+          description: Plot,
+          imgUrl: movieImg,
+          imdbUrl: `https://www.imdb.com/title/${imdbID}`,
+          imdbId: imdbID,
+        };
+
+        setIsMovieExist(false);
+        setMovie(newMovieInList);
+      } else {
+        setIsMovieExist(true);
+        setMovie(null);
+      }
+    } catch {
+      throw new Error();
+    }
+  };
+
+  const handleSearch = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    setIsSearch(true);
+    getMovieFromServer(query);
+  };
+
+  const checkMovie = (newMovie: Movie) => {
+    return movies.some(film => film.imdbId === newMovie.imdbId);
+  };
+
+  const addingMovie = () => {
+    if (movie) {
+      if (!checkMovie(movie)) {
+        setMovies([...movies, movie]);
+      }
+    }
+
+    setMovie(null);
+    setQuery('');
+  };
+
   const contextValue = {
     movies,
-    setMovies,
     query,
-    setQuery,
     movie,
-    setMovie,
     isMovieExist,
-    setIsMovieExist,
     isSearch,
-    setIsSearch,
+
+    handleInput,
+    getMovieFromServer,
+    handleSearch,
+    addingMovie,
   };
 
   return (
